renaming sir_base to sir
