+++ examples/sir.py
